[usm] Move `telemetry` package to `protocols` namespace
