+++ pkg/network/protocols/telemetry/delta.go
