+++ pkg/network/protocols/telemetry/registry.go
