+++ pkg/network/protocols/telemetry/utils.go
